markdown: adds cleanup for excess newlines from parsers
Some block parsers add new lines before and after its content to ensure
they are split from surrounding content in markdown. This leads to
an excess of newlines whenever two of this type of blocks are next to
each other.

A cleanup function handles the final parsing result and ensures only a
maximum of two newlines are used for block separation.